moving test format into it's own file
